Consolidate .netcoreapp.cs test files in System.Runtime.Extensions (#42184)
* Consolidate .netcoreapp.cs files because System.Runtime.Extensions project is no longer cross-compiled

* Fix AppDomainTests
