+++ src/System.Runtime.Extensions/tests/System.Runtime.Extensions.Tests.csproj
@@ -6,24 +6,14 @@
     <IncludeRemoteExecutor>true</IncludeRemoteExecutor>
     <Configurations>netcoreapp-Unix-Debug;netcoreapp-Unix-Release;netcoreapp-Windows_NT-Debug;netcoreapp-Windows_NT-Release</Configurations>
   </PropertyGroup>
-  <ItemGroup Condition="'$(TargetsNetCoreApp)' == 'true'">
-    <Compile Include="System\AppDomainTests.netcoreapp.cs" />
-    <Compile Include="System\Random.netcoreapp.cs" />
-    <Compile Include="System\IO\Path.IsPathFullyQualified.netcoreapp.cs" />
-    <Compile Include="System\IO\PathTestsBase.netcoreapp.cs" />
-    <Compile Include="System\IO\PathTests_Windows.netcoreapp.cs" />
-    <Compile Include="System\IO\PathTests.netcoreapp.cs" />
-    <Compile Include="System\IO\PathTests_Unix.cs" />
-    <Compile Include="System\IO\PathTests_Join.netcoreapp.cs" />
-    <Compile Include="System\Runtime\ProfileOptimization.netcoreapp.cs" />
-    <Compile Include="System\Runtime\CompilerServices\SwitchExpressionExceptionTests.cs" />
-  </ItemGroup>
   <ItemGroup>
     <Compile Include="System\ApplicationIdTests.cs" />
     <Compile Include="System\Convert.cs" />
     <Compile Include="System\EnvironmentTests.cs" />
     <Compile Include="System\IO\PathTests.cs" />
     <Compile Include="System\IO\PathTests_Windows.cs" />
+    <Compile Include="System\IO\PathTests_Unix.cs" />
+    <Compile Include="System\IO\PathTests_Join.cs" />
     <Compile Include="System\OperatingSystemTests.cs" />
     <Compile Include="System\Runtime\Versioning\VersioningHelperTests.cs" />
     <Compile Include="System\AppDomainTests.cs" />
@@ -31,14 +21,10 @@
     <Compile Include="System\MarshalByRefObjectTest.cs" />
     <Compile Include="TestHelpers.cs" />
     <Compile Include="System\BitConverterSpan.cs" />
-    <Compile Include="System\BitConverter.netcoreapp.cs" />
-    <Compile Include="System\Convert.netcoreapp.cs" />
     <Compile Include="System\IO\Path.GetRelativePath.cs" />
-    <Compile Include="System\MathTests.netcoreapp.cs" />
-    <Compile Include="System\MathF.netcoreapp.cs" />
+    <Compile Include="System\MathF.cs" />
     <Compile Include="System\Numerics\BitOperationsTests.cs" />
-    <Compile Include="System\StringComparer.netcoreapp.cs" />
-    <Compile Include="System\UnloadingAndProcessExitTests.netcoreapp.cs" />
+    <Compile Include="System\UnloadingAndProcessExitTests.cs" />
     <Compile Include="System\IO\Path.IsPathFullyQualified.cs" />
     <Compile Include="System\BitConverterArray.cs" />
     <Compile Include="System\BitConverterBase.cs" />
@@ -84,6 +70,9 @@
     <Compile Include="System\Progress.cs" />
     <Compile Include="System\Random.cs" />
     <Compile Include="System\StringComparer.cs" />
+    <Compile Include="System\Runtime\ProfileOptimization.cs" />
+    <Compile Include="System\Runtime\CompilerServices\SwitchExpressionExceptionTests.cs" />
+
     <Compile Include="$(CommonPath)\Interop\Unix\System.Native\Interop.GetNodeName.cs" Condition="'$(TargetsUnix)' == 'true'">
       <Link>Common\Interop\Unix\System.Native\Interop.GetNodeName.cs</Link>
     </Compile>
